+++ src/main/java/com/jsorrell/carpetskyadditions/mixin/CreateWorldScreenMixin.java
@@ -2,6 +2,7 @@ package com.jsorrell.carpetskyadditions.mixin;
 
 import com.jsorrell.carpetskyadditions.config.SkyAdditionsConfig;
 import com.jsorrell.carpetskyadditions.gen.SkyAdditionsWorldPresets;
+import com.jsorrell.carpetskyadditions.helpers.DataConfigurationHelper;
 import me.shedaniel.autoconfig.AutoConfig;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
@@ -11,14 +12,13 @@ import net.minecraft.client.gui.screen.world.CreateWorldScreen;
 import net.minecraft.registry.DynamicRegistryManager;
 import net.minecraft.registry.RegistryKey;
 import net.minecraft.registry.RegistryKeys;
+import net.minecraft.resource.DataConfiguration;
 import net.minecraft.world.dimension.DimensionOptionsRegistryHolder;
 import net.minecraft.world.gen.WorldPreset;
 import net.minecraft.world.gen.WorldPresets;
 import org.objectweb.asm.Opcodes;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.Redirect;
+import org.spongepowered.asm.mixin.injection.*;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Environment(value = EnvType.CLIENT)
@@ -27,7 +27,7 @@ public class CreateWorldScreenMixin {
     // Try to reload config whenever Create New World is clicked
     @Inject(
             method = "create(Lnet/minecraft/client/MinecraftClient;Lnet/minecraft/client/gui/screen/Screen;)V",
-            at = @At(value = "HEAD"))
+            at = @At("HEAD"))
     private static void loadConfigFromFile(MinecraftClient client, Screen parent, CallbackInfo ci) {
         AutoConfig.getConfigHolder(SkyAdditionsConfig.class).load();
     }
@@ -64,5 +64,16 @@ public class CreateWorldScreenMixin {
         } else {
             return WorldPresets.createDemoOptions(drm);
         }
+    }
+
+    @ModifyArg(
+            method = "create(Lnet/minecraft/client/MinecraftClient;Lnet/minecraft/client/gui/screen/Screen;)V",
+            at =
+                    @At(
+                            value = "INVOKE",
+                            target =
+                                    "Lnet/minecraft/client/gui/screen/world/CreateWorldScreen;createServerConfig(Lnet/minecraft/resource/ResourcePackManager;Lnet/minecraft/resource/DataConfiguration;)Lnet/minecraft/server/SaveLoading$ServerConfig;"))
+    private static DataConfiguration enableSkyAdditionsDatapacks(DataConfiguration dc) {
+        return DataConfigurationHelper.updateDataConfiguration(dc);
     }
 }